+++ lib/slack/sends.ex
@@ -128,10 +128,14 @@ defmodule Slack.Sends do
   defp open_im_channel(token, user_id, on_success, on_error) do
     im_open =
       with url <- Application.get_env(:slack, :url, "https://slack.com") <> "/api/im.open",
-           headers <- {:form, [token: token, user: user_id]},
-           options <-
-             Application.get_env(:slack, :web_http_client_opts, [{"Content-Type", "utf-8"}]) do
-        HTTPoison.post(url, headers, options)
+           options <- Application.get_env(:slack, :web_http_client_opts, []) do
+        Req.post(
+          url,
+          Keyword.merge(options,
+            form: [token: token, user: user_id],
+            decode_body: false
+          )
+        )
       end
 
     case im_open do